# [feature] Optimize Contributors Page for Large Projects [OSF-8536]