#### Purpose
- Speeds up the contributors page for large and deeply nested projects. 

#### Changes
- Greatly reduces the number of queries made by the `_get_children` and `node_child_tree` v1 views by using subqueries, annotations, and django-include. 
- Consolidates some serialization logic so that expensive permissions checks are executed once instead of 3-4 times per node.

#### Side Effects
- Faster! Otherwise, none expected.

#### Stats 🐱 
- Local response times and num queries when testing against a QA project (23rez) with ~100 children:
```
BEFORE
/api/v1/project/<guid>/_get_editable_children/:
    response time: 
        ~3s (admin on root, r+w on children)
        ~23s (admin on root, admin on children)
    num queries:
        43 (admin on root, r+w on children)
        373 (admin on root, admin on children)
/api/v1/project/<guid>/_tree response time:
    response time: 
        ~38s (admin on root, admin on children)
        ~52s (admin on root, r+w on children) 
    num queries:
        1267 -- not a typo, it's really 1267 (admin on root, admin on children) 
        2170 -- I *reallly* wish I was kidding (admin on root, r+w on children)

AFTER
/api/v1/project/<guid>/_get_editable_children/:
    response time:
        ~1s (admin on root, r+w on children)
        ~3s (admin on root, admin on children)
    num queries:
        1 (admin on root, admin on children) 
        1 (admin on root, r+w on children) 

/api/v1/project/<guid>/_tree response time:
    response time:
        ~3s (admin on root, admin on children)
        ~3s (admin on root, r+w on children) 
    num queries:
        12 (admin on root, admin on children) 
        12 (admin on root, r+w on children)
```

#### QA Notes
- The add VOL modal (specifically the "which components would you like to associate with this link?" bit) and the add contributor modal (specifically the "select components" page) should be tested, especially with complicated parent-child permission combinations. 
- Note that the VOL modal will no longer have ordering (if it did before?), but children will still appear nested (indented) below their parents. 
- Note that the components in the "select components" modal will now be ordered by number of children.

#### Ticket
- [OSF-8536](https://openscience.atlassian.net/browse/OSF-8536)