@sloria 2017-10-26
Merged into 0.125.0